# use setters for attributes to make setting faster