@balvig 2017-06-26
@samandmoore, I'm so glad you found this project! 😅  🙏 

Adding the setter in addition to the getter definitely makes sense, thanks for this optimization! 👍  

The `attributes` method was originally added to protect the app against APIs that would sometimes leave out attributes when blank, so I was initially mostly concerned with the `getting` 😅 

> I think we probably want to extend this behavior to include predicate methods and associations. What do you think?

Yes, I think you're right. I can't think of a reason why `attributes :foo` shouldn't get the speed boost for `foo?`, and I guess adding an association with a `has_many` macro etc should have the same effect, ie methods could get defined so we don't have to rely on `method_missing` 🤔 

> I'm curious how you feel about ActiveModel::AttributeMethods. I am wondering if it would be worth investigating migrating the AttributeAssignment mixin to use that approach.

That does indeed look like it might be useful! I can't say I'm familiar enough with it to know whether it would "fit right in", but would certainly be interested to see if that could help out with some of the heavy lifting. I've been meaning to look into the new attributes API as well, whether to see if that might gives us some of the functionality for free 🤔